requirements.txt
# Core libraries
opencv-python==4.5.4.58
numpy==1.21.2

# Web application (Flask)
flask==2.1.0
gunicorn==20.1.0

# Pre-trained models
torch==1.10.0
torchvision==0.11.1

# DALL-E (optional, if using DALL-E)
DALL-E==0.1


# Miscellaneous
Pillow==8.4.0
tqdm==4.62.3
